+++ images/common/harness/run_harness.rb
@@ -9,7 +9,7 @@ class RunHarness < BaseHarness
     ENV['PRG']
   end
 
-  def do_validate_other_env
+  def validate_other_env
     abort 'No program path given.' if !program_path
   end
 
@@ -24,10 +24,6 @@ class RunHarness < BaseHarness
   def validate_input_json
     abort 'No program given.' if !program
     abort 'No stdin given.' if !stdin
-  end
-
-  def do_stdin
-    stdin
   end
 
 end