+++ sandboxes/common/test/lib/input-json-for-run-interpreted.bash
@@ -4,5 +4,7 @@ function input_json {
   set -e
   var_is_set interpreted_program
   : ${stdin:=''}
-  echo "{ \"stdin\":\"$stdin\" }" | jshon -s "$interpreted_program" -i interpretedProgram
+  echo '{}' | \
+    jshon -s "$stdin" -i stdin | \
+    jshon -s "$interpreted_program" -i interpretedProgram
 }